move statement to sql.ast
