+++ fe/fe-core/src/main/cup/sql_parser.cup
@@ -28,10 +28,10 @@ import com.starrocks.analysis.AlterDatabaseQuotaStmt.QuotaType;
 import com.starrocks.analysis.SetOperationStmt.Qualifier;
 import com.starrocks.analysis.SetOperationStmt.Operation;
 import com.starrocks.analysis.SetOperationStmt.SetOperand;
-import com.starrocks.analysis.CreateResourceGroupStmt;
-import com.starrocks.analysis.AlterResourceGroupStmt;
-import com.starrocks.analysis.DropResourceGroupStmt;
-import com.starrocks.analysis.ShowResourceGroupStmt;
+import com.starrocks.sql.ast.CreateResourceGroupStmt;
+import com.starrocks.sql.ast.AlterResourceGroupStmt;
+import com.starrocks.sql.ast.DropResourceGroupStmt;
+import com.starrocks.sql.ast.ShowResourceGroupStmt;
 import com.starrocks.catalog.AccessPrivilege;
 import com.starrocks.catalog.AggregateType;
 import com.starrocks.catalog.ArrayType;
