+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/DropResourceGroupStmt.java
@@ -1,7 +1,7 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
-package com.starrocks.analysis;
+package com.starrocks.sql.ast;
 
-import com.starrocks.sql.ast.AstVisitor;
+import com.starrocks.analysis.DdlStmt;
 
 // Drop ResourceGroup specified by name
 // DROP RESOURCE GROUP <name>
