+++ fe/fe-core/src/main/java/com/starrocks/sql/ast/ShowResourceGroupStmt.java
@@ -1,10 +1,10 @@
 // This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
 
-package com.starrocks.analysis;
+package com.starrocks.sql.ast;
 
+import com.starrocks.analysis.ShowStmt;
 import com.starrocks.catalog.ResourceGroup;
 import com.starrocks.qe.ShowResultSetMetaData;
-import com.starrocks.sql.ast.AstVisitor;
 
 // Show ResourceGroups
 // 1. Show ResourceGroup specified by name
